+++ tests/commands/delete/test_delete_cmd.py
@@ -1,12 +1,9 @@
 # -*- coding: utf-8 -*-
-
-from pymongo import ASCENDING, IndexModel
-
 from scout.commands import cli
 from scout.server.extensions import store
 
-VARIANTS_QUERY = {"rank_score": {"$lt": 15}}
-RANK_THRESHOLD = 15
+VARIANTS_QUERY = {"rank_score": {"$lt": 0}}
+RANK_THRESHOLD = 0
 VARIANTS_THRESHOLD = 10
 
 
@@ -55,14 +52,22 @@ def test_delete_variants_dry_run(mock_app, case_obj, user_obj):
 def test_delete_variants(mock_app, case_obj, user_obj):
     """Test deleting variants using the delete variants command line"""
 
-    # Given a database with SNV variants
+    # Given a case with with SNV variants
     runner = mock_app.test_cli_runner()
     result = runner.invoke(
-        cli, ["load", "variants", case_obj["_id"], "--snv", "--rank-treshold", 5]
+        cli, ["load", "variants", "--snv", "--rank-treshold", 0, case_obj["_id"]]
     )
     assert result.exit_code == 0
-    n_initial_vars = sum(1 for _ in store.variant_collection.find())
-    n_variants_to_exclude = store.variant_collection.count_documents(VARIANTS_QUERY)
+    nr_snvs = sum(1 for _ in store.variant_collection.find())
+
+    # AND WTS outliers
+    result = runner.invoke(
+        cli, ["load", "variants", "--outlier-research", case_obj["_id"], "--force"]
+    )
+    assert result.exit_code == 0
+    nr_outliers = sum(1 for _ in store.omics_variant_collection.find())
+
+    n_initial_vars = nr_snvs + nr_outliers
 
     # Then the function that delete variants should run without error
     cmd_params = [
@@ -72,6 +77,8 @@ def test_delete_variants(mock_app, case_obj, user_obj):
         user_obj["email"],
         "--status",
         "inactive",
+        "--keep-ctg",
+        "outlier",
         "--older-than",
         2,
         "--analysis-type",
@@ -82,20 +89,65 @@ def test_delete_variants(mock_app, case_obj, user_obj):
         VARIANTS_THRESHOLD,
     ]
     result = runner.invoke(cli, cmd_params, input="y")
+
     assert result.exit_code == 0
     assert "estimated deleted variants" not in result.output
     assert "Estimated space freed" in result.output
     # variants should be deleted
     n_current_vars = sum(1 for _ in store.variant_collection.find())
     assert n_current_vars < n_initial_vars
-    assert n_current_vars + n_variants_to_exclude == n_initial_vars
+
     # and a relative event should be created
     event = store.event_collection.find_one({"verb": "remove_variants"})
     assert event["case"] == case_obj["_id"]
     assert (
         event["content"]
-        == f"Rank-score threshold:{RANK_THRESHOLD}, case n. variants threshold:{VARIANTS_THRESHOLD}."
+        == f"Rank-score threshold:0, case n. variants threshold:{VARIANTS_THRESHOLD}."
     )
+    # SNV variants should be gone
+    assert sum(1 for _ in store.variant_collection.find()) == 0
+    # WHILE outliers should still be available
+    assert sum(1 for _ in store.omics_variant_collection.find()) == nr_outliers
+
+
+def test_delete_outlier_variants(mock_app, case_obj, user_obj):
+    """Test the delete variants command's ability to remove omics variants."""
+
+    # Given a case with with (research) outlier variants
+    runner = mock_app.test_cli_runner()
+    result = runner.invoke(
+        cli, ["load", "variants", "--outlier-research", case_obj["_id"], "--force"]
+    )
+    assert result.exit_code == 0
+    n_initial_vars = sum(1 for _ in store.omics_variant_collection.find())
+    assert n_initial_vars
+    n_variants_to_delete = store.omics_variant_collection.count_documents({})
+    assert n_variants_to_delete
+
+    # WHEN variants are removed using the command line
+    cmd_params = [
+        "delete",
+        "variants",
+        "-u",
+        user_obj["email"],
+        "--rank-threshold",
+        0,
+        "--rm-ctg",
+        "outlier",
+    ]
+    result = runner.invoke(cli, cmd_params, input="y")
+    assert result.exit_code == 0
+    assert "estimated deleted variants" not in result.output
+    assert "Estimated space freed" in result.output
+
+    # THEN the variants should be gone
+    n_current_vars = sum(1 for _ in store.variant_collection.find())
+    assert n_current_vars == 0
+    assert n_current_vars + n_variants_to_delete == n_initial_vars
+    # and a relative event should be created
+    event = store.event_collection.find_one({"verb": "remove_variants"})
+    assert event["case"] == case_obj["_id"]
+    assert "Rank-score threshold:0" in event["content"]
 
 
 def test_delete_panel_non_existing(empty_mock_app, testpanel_obj):